requirements.txt
datasets==2.20.0
keras==3.4.1
kfp==2.8.0
kfp_pipeline_spec==0.3.0
numpy==1.24.1
pandas==1.5.2
scikit_learn==1.3.2
torch==2.4.0
transformers==4.42.4
flake8==6.0.0
isort==5.11.4
mypy==0.991
pydocstyle==6.3.0
torchmetrics==1.4.0
tensorflow==2.17.0
pre-commit==3.8.0
google-cloud-aiplatform==1.62.0
gcsfs==2023.12.1